+++ __unported__/magentoerpconnect/tests/test_synchronization.py
@@ -28,8 +28,10 @@ from openerp.addons.magentoerpconnect.unit.import_synchronizer import (
 from openerp.addons.connector.session import ConnectorSession
 import openerp.tests.common as common
 from .common import (mock_api,
-                     mock_urlopen_image)
+                     mock_urlopen_image,
+                     MagentoHelper)
 from .test_data import magento_base_responses
+
 
 DB = common.DB
 ADMIN_USER_ID = common.ADMIN_USER_ID
@@ -46,6 +48,7 @@ class SetUpMagentoBase(common.TransactionCase):
         super(SetUpMagentoBase, self).setUp()
         self.backend_model = self.registry('magento.backend')
         self.session = ConnectorSession(self.cr, self.uid)
+        self.session.context['__test_no_commit'] = True
         data_model = self.registry('ir.model.data')
         self.get_ref = partial(data_model.get_object_reference,
                                self.cr, self.uid)
@@ -77,6 +80,9 @@ class SetUpMagentoBase(common.TransactionCase):
                  'import_rule': 'always',
                  'days_before_cancel': 0,
                  'journal_id': journal_id})
+
+    def get_magento_helper(self, model_name):
+        return MagentoHelper(self.cr, self.registry, model_name)
 
 
 class TestBaseMagento(SetUpMagentoBase):
@@ -308,8 +314,13 @@ class TestImportMagento(SetUpMagentoSynchronized):
     def test_34_import_sale_order_with_taxes_included(self):
         """ Import sale order with taxes included """
         backend_id = self.backend_id
-        self.backend_model.write(self.cr, self.uid, self.backend_id,
-                                 {'catalog_price_tax_included': True})
+        storeview_model = self.registry('magento.storeview')
+        storeview_id = storeview_model.search(
+            self.cr, self.uid,
+            [('backend_id', '=', backend_id),
+             ('magento_id', '=', '1')])
+        storeview_model.write(self.cr, self.uid, storeview_id,
+                              {'catalog_price_tax_included': True})
         with mock_api(magento_base_responses):
             with mock_urlopen_image():
                 import_record(self.session,
@@ -333,14 +344,19 @@ class TestImportMagento(SetUpMagentoSynchronized):
         # 97.5 is the amount_total if connector takes correctly included
         # tax prices.
         self.assertEqual(amount_total, 97.5000)
-        self.backend_model.write(self.cr, self.uid, self.backend_id,
-                                 {'catalog_price_tax_included': False})
+        storeview_model.write(self.cr, self.uid, storeview_id,
+                              {'catalog_price_tax_included': False})
 
     def test_35_import_sale_order_with_discount(self):
         """ Import sale order with discounts"""
         backend_id = self.backend_id
-        self.backend_model.write(self.cr, self.uid, self.backend_id,
-                                 {'catalog_price_tax_included': True})
+        storeview_model = self.registry('magento.storeview')
+        storeview_id = storeview_model.search(
+            self.cr, self.uid,
+            [('backend_id', '=', backend_id),
+             ('magento_id', '=', '2')])
+        storeview_model.write(self.cr, self.uid, storeview_id,
+                              {'catalog_price_tax_included': True})
         with mock_api(magento_base_responses):
             with mock_urlopen_image():
                 import_record(self.session,
@@ -371,5 +387,5 @@ class TestImportMagento(SetUpMagentoSynchronized):
                 self.fail('encountered unexpected sale '
                           'order line %s' % line.name)
 
-        self.backend_model.write(self.cr, self.uid, self.backend_id,
-                                 {'catalog_price_tax_included': False})
+        storeview_model.write(self.cr, self.uid, storeview_id,
+                              {'catalog_price_tax_included': False})