+++ __unported__/magentoerpconnect_order_comment/tests/test_synchronization.py
@@ -158,4 +158,5 @@ class TestMagentoSaleCommentExport(SetUpMagentoWithSaleOrder):
             method, (mag_order_id, state, comment, notif) = calls_done[0]
             self.assertEqual(method, 'sales_order.addComment')
             self.assertEqual(mag_order_id, '900000691')
-            self.assertEqual(comment, 'Test me I famous')
+            # the comment is in <p> so BeautifulSoup adds a final \n
+            self.assertEqual(comment.rstrip('\n'), 'Test me I famous')