+++ __unported__/magentoerpconnect_pricing/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-import connector
-import magento_model
-import product
-import sale
+from . import connector
+from . import magento_model
+from . import product
+from . import sale